+++ core/src/test/scala/org/apache/spark/DistributedSuite.scala
@@ -23,7 +23,6 @@ import org.scalatest.time.{Millis, Span}
 
 import org.apache.spark.internal.config
 import org.apache.spark.internal.config.Tests._
-import org.apache.spark.network.SyncBlockTransferClient
 import org.apache.spark.security.EncryptionFunSuite
 import org.apache.spark.storage.{RDDBlockId, StorageLevel}
 import org.apache.spark.util.io.ChunkedByteBuffer
@@ -186,18 +185,13 @@ class DistributedSuite extends SparkFunSuite with Matchers with LocalSparkContex
     val blockId = blockIds(0)
     val blockManager = SparkEnv.get.blockManager
     val blockTransfer = blockManager.blockTransferService
-    val blockTransferClientSync = new SyncBlockTransferClient(blockTransfer)
     val serializerManager = SparkEnv.get.serializerManager
     val locations = blockManager.master.getLocations(blockId)
     assert(locations.size === storageLevel.replication,
       s"; got ${locations.size} replicas instead of ${storageLevel.replication}")
     locations.foreach { cmId =>
-      val bytes = blockTransferClientSync.fetchBlockSync(
-        cmId.host,
-        cmId.port,
-        cmId.executorId,
-        blockId.toString,
-        null)
+      val bytes = blockTransfer.fetchBlockSync(cmId.host, cmId.port, cmId.executorId,
+        blockId.toString, null)
       val deserialized = serializerManager.dataDeserializeStream(blockId,
         new ChunkedByteBuffer(bytes.nioByteBuffer()).toInputStream())(data.elementClassTag).toList
       assert(deserialized === (1 to 100).toList)
