+++ core/src/test/scala/org/apache/spark/network/BlockTransferServiceSuite.scala
@@ -93,11 +93,10 @@ class BlockTransferServiceSuite extends SparkFunSuite with TimeLimits {
       }
     }
 
-    val blockTransferClientSync = new SyncBlockTransferClient(blockTransferService)
     val e = intercept[SparkException] {
       failAfter(10.seconds) {
-        blockTransferClientSync.fetchBlockSync("localhost-unused", 0, "exec-id-unused",
-          "block-id-unused", null)
+        blockTransferService.fetchBlockSync(
+          "localhost-unused", 0, "exec-id-unused", "block-id-unused", null)
       }
     }
     assert(e.getCause.isInstanceOf[IllegalArgumentException])
